separated deletetask.js
