+++ src/modules/deletetask.js
@@ -1,0 +1,13 @@
+import { updateTaskIndices, updateLocalStorage } from "./addtask";
+
+function deleteTask(listItem, tasks) {
+    const taskList = listItem.parentNode;
+    const taskIndex = Array.from(taskList.children).indexOf(listItem);
+    taskList.removeChild(listItem);
+  
+    tasks.splice(taskIndex, 1);
+  
+    updateTaskIndices(tasks);
+    updateLocalStorage(tasks);
+  }
+export default deleteTask;
